+++ packages/core/src/audio/use-frame-for-volume-prop.tsx
@@ -1,12 +1,5 @@
-import {useContext} from 'react';
-import {SequenceContext} from '../sequencing';
 import {useCurrentFrame} from '../use-frame';
-
-export const useMediaStartsAt = () => {
-	const parentSequence = useContext(SequenceContext);
-	const startsAt = Math.min(0, parentSequence?.relativeFrom ?? 0);
-	return startsAt;
-};
+import {useMediaStartsAt} from './use-media-starts-at';
 
 /**
  * When passing a function as the prop for `volume`,
